refactor: 보드 확장 여부 제거에 따른 BoardStructStrategy 제거
