+++ src/main/java/team/janggi/domain/board/BoardStatus.java
@@ -1,6 +1,8 @@
-package team.janggi.domain;
+package team.janggi.domain.board;
 
 import java.util.Map;
+import team.janggi.domain.Position;
+import team.janggi.domain.Team;
 import team.janggi.domain.piece.Piece;
 
 public interface BoardStatus {
