--- src/main/java/team/janggi/domain/strategy/boardstruct/BoardStructStrategy.java
@@ -1,9 +1,0 @@
-package team.janggi.domain.strategy.boardstruct;
-
-import team.janggi.domain.BoardStatus;
-
-public interface BoardStructStrategy {
-
-    void initBoardStatus(BoardStatus status);
-
-}
