+++ src/main/java/team/janggi/domain/strategy/layout/LayoutStrategy.java
@@ -1,6 +1,6 @@
 package team.janggi.domain.strategy.layout;
 
-import team.janggi.domain.BoardStatus;
+import team.janggi.domain.board.BoardStatus;
 
 public interface LayoutStrategy {
