+++ src/test/java/team/janggi/domain/piece/CannonTest.java
@@ -4,12 +4,12 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import team.janggi.domain.BoardStatus;
+import team.janggi.domain.board.BoardStatus;
 import team.janggi.domain.EmptyLayoutStrategy;
 import team.janggi.domain.LocalMemoryBoardStatus;
 import team.janggi.domain.Position;
 import team.janggi.domain.Team;
-import team.janggi.domain.strategy.boardstruct.NormalBoardStrategy;
+import team.janggi.domain.board.BoardFactory;
 
 public class CannonTest {
 
@@ -17,7 +17,7 @@ public class CannonTest {
 
     @BeforeEach
     void setUp() {
-        new NormalBoardStrategy(EmptyLayoutStrategy.instance).initBoardStatus(boardStatus);
+        new BoardFactory(EmptyLayoutStrategy.instance).initBoardStatus(boardStatus);
     }
 
     @ParameterizedTest
